Move ProjectType to its own file.
