+++ physionet-django/project/models.py
@@ -5,6 +5,7 @@ from project.modelcomponents.coreproject import *
 from project.modelcomponents.fields import *
 from project.modelcomponents.legacy import *
 from project.modelcomponents.metadata import *
+from project.modelcomponents.projecttype import *
 from project.modelcomponents.publishedproject import *
 from project.modelcomponents.storage import *
 from project.modelcomponents.submission import *
